Split tierSelect into components
