+++ src/app/upgrade/_components/TierSelect.tsx
@@ -5,120 +5,10 @@ import { Switch } from '@headlessui/react'
 import { CheckCircleIcon } from '@heroicons/react/20/solid'
 import { api } from "~/trpc/react"
 
-import { useRouter } from "next/navigation"
+import { useRouter, useSearchParams } from "next/navigation"
 import { useSession, signIn } from "next-auth/react"
-
-type price = {
-  monthly: string,
-  annually_m: string,
-  annually_y: string
-}
-
-type priceId = {
-  monthly: string,
-  annually: string,
-}
-
-interface tier {
-  name: string,
-  id: string,
-  href: string,
-  price: price,
-  priceId?: priceId,
-  description: string
-  features: string[]
-}
-
-const tiers : tier[] = [
-  {
-    name: '20x30 All Purpose',
-    id: 'tier-pleb',
-    href: '#',
-    price: { monthly: 'Free', annually_m: 'Free', annually_y: 'Free' },
-    description: 'Everything necessary to get started. We can be nice if we wanted to.',
-    features: ['Free 15 GB storage just for you ❤️'],
-  },
-  {
-    name: '40x60 Heavy Duty',
-    id: 'tier-normal',
-    href: '#',
-    price: { monthly: '$1.99', annually_m: '$1.49', annually_y: '$17.99' },
-    priceId: { monthly: 'price_1O95MAIEdHdbj4cnIlR0sIgM', annually: 'price_1O95TNIEdHdbj4cn850p9f5d' },
-    description: 'When you need some serious storage for them 🐈‍⬛ pics.',
-    features: [
-      '100 GB of on-demand storage',
-      'We will actually answer your emails'
-    ],
-  },
-  {
-    name: '80x120 Industrial Grade',
-    id: 'tier-whale',
-    href: '#',
-    price: { monthly: '$5.99', annually_m: '$4.99', annually_y: '$59.99' },
-    priceId: { monthly: 'price_1O98rkIEdHdbj4cndIUjsLeR', annually: 'price_1O98rkIEdHdbj4cnQRHrt7DY' },
-    description: 'You are an instagram model or something, idk. You do you.',
-    features: [
-      'A massive 1 TB of storage',
-      'You might be able to call us also'
-
-    ],
-  },
-]
-
-function PricePerMonth(props : {tier : tier})
-{
-  if (props.tier.id == undefined || props.tier.price.monthly == undefined)
-    return
-    
-  else if (props.tier.id != 'tier-pleb') {
-    return (
-      <div>
-        <p className="mt-6 pb-9 flex items-baseline gap-x-1">
-          <span className="text-5xl font-bold tracking-tight text-gray-900">{props.tier.price.monthly}</span> 
-          <span className="text-sm font-semibold leading-6 text-gray-600">/month</span>
-        </p>
-      </div>
-      );
-  }
-  else {
-    return( 
-      <div>
-        <p className="mt-6 pb-9 flex items-baseline gap-x-1">
-          <span className="text-5xl font-bold tracking-tight text-gray-900">{props.tier.price.monthly}</span>
-        </p>
-      </div>
-    );
-  }
-}
-
-function PricePerAnnum(props : {tier : tier})
-{
-  if (props.tier.id == undefined || props.tier.price.annually_y == undefined || props.tier.price.annually_m == undefined)
-    return
-    
-  else if (props.tier.id != 'tier-pleb') {
-    return (
-      <div>
-        <p className="mt-6 flex items-baseline gap-x-1">
-          <span className="text-5xl font-bold tracking-tight text-gray-900">{props.tier.price.annually_m}</span> 
-          <span className="text-sm font-semibold leading-6 text-gray-600">/month</span>
-        </p>
-        <p className="mt-3 text-sm leading-6 text-gray-500">
-          {props.tier.price.annually_y} billed per year
-        </p>
-      </div>
-      );
-  }
-  else {
-    return( 
-      <div>
-        <p className="mt-6 flex items-baseline gap-x-1">
-          <span className="text-5xl pb-9 font-bold tracking-tight text-gray-900">{props.tier.price.monthly}</span> 
-        </p>
-      </div>
-    );
-  }
-}
+import { tier,tiers } from "./Tiers"
+import { PricePerAnnum, PricePerMonth } from "./TierFunctions"
 
 export function TierSelect() {
   const { data: session } = useSession();
@@ -158,9 +48,10 @@ export function TierSelect() {
     }
     else
       signIn()
-      
-    
   }
+
+  const searchParams = useSearchParams();
+  const sessionId = searchParams.get('sessionId');
 
   return (
     <div className={`flex justify-center w-full`}>
